+++ jni/game/entities/KosmoObject.h
@@ -19,7 +19,7 @@ public:
 
     KosmoObject(structuretype collisionStructureType, mjResourceManager* resourceManager, Level* levelData);
     KosmoObject(mjResourceManager* resourceManager, Level* levelData);
-    void Update(float t_elapsed) override;
+    void Update(double t_elapsed) override;
     virtual void StartTransitionForTravel(){}
     virtual void EndTransitionForTravel(){}
 };
